+++ CHANGELOG.md
@@ -63,6 +63,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Health tracking for NPCs in unloaded chunks.
 - More snow covered variants of grass and flowers
 - A downed state, at the moment of death humanoids survive with 1 hp, and have to be helped up.
+- State of the window is saved between game launches.
 
 ### Changed